+++ src/routes/dashboard.jsx
@@ -1,22 +1,25 @@
 import Navbar from "@app/components/site/Header.jsx";
-import Container from "@ui/Container.jsx";
-import { createFileRoute, Outlet } from "@tanstack/react-router";
-import { cn } from "@app/lib/utils.js";
-import PatientDashboard from "@components/site/patient/PatientDashboard.jsx";
 import DoctorDashboard from "@components/site/doctor/DoctorDashboard.jsx";
 import FinanceDashboard from "@components/site/finance/FinanceDashboard.jsx";
+import PatientDashboard from "@components/site/patient/PatientDashboard.jsx";
+import { cn } from "@lib/utils.js";
+import { createFileRoute } from "@tanstack/react-router";
+import Container from "@ui/Container.jsx";
+import { useSelector } from "react-redux";
 
 export const Route = createFileRoute("/dashboard")({
   component: RouteComponent,
 });
 
 function RouteComponent() {
-  const role = "doctor"; // This would be fetched from user context/auth in a real app
+  const { role, name } = useSelector((state) => state.user.user);
   return (
     <Container className={cn("mx-auto max-w-7xl px-4 py-6")}>
       <Navbar />
       <Container>
-        <h1 className={cn("text-3xl font-semibold")}>Welcome, (name)</h1>
+        <h1 className={cn("text-3xl font-semibold capitalize")}>
+          welcome, {name}
+        </h1>
       </Container>
       {role === "patient" && <PatientDashboard />}
       {role === "doctor" && <DoctorDashboard />}