+++ src/store/store.js
@@ -1,9 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
-import themeReducer from "@store/slices/theme.js";
+import themeReducer from "@store/slices/theme.slice.js";
+import userReducer from "@store/slices/user.slice.js";
 
 const store = configureStore({
   reducer: {
     theme: themeReducer,
+    user: userReducer,
   },
 });