+++ src/ui/Providers.jsx
@@ -1,8 +1,13 @@
 import { store } from "@store/store.js";
 import { Provider } from "react-redux";
+import ProtectedRoute from "./ProtectedRoute.jsx";
 
 const Providers = ({ children }) => {
-  return <Provider store={store}>{children}</Provider>;
+  return (
+    <Provider store={store}>
+      <ProtectedRoute>{children}</ProtectedRoute>
+    </Provider>
+  );
 };
 
 export default Providers;